requirements.txt
python-telegram-bot==21.6
google-generativeai==0.8.3
pytrends==4.9.2
feedparser==6.0.11
requests==2.32.3
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
